Move request_reference into the top level namespace
At the moment, we're not calling the request_reference service
in the support UI when a support user reinstates a
reference request. We should be. Due to this, the service
needs to be moved to the top level namespace.
