+++ app/controllers/support_interface/references_controller.rb
@@ -15,7 +15,7 @@ module SupportInterface
     def reinstate; end
 
     def confirm_reinstate
-      @reference.update!(feedback_status: 'feedback_requested')
+      RequestReference.new.call(@reference)
       flash[:success] = 'Reference was reinstated'
       redirect_to support_interface_application_form_path(@reference.application_form)
     end
